remove the shared package for now until a proper pnpm monorepo setup
