+++ api/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.5.0-alpha.2",
+  "version": "0.7.0-alpha.4",
   "name": "@fiberplane/studio",
   "description": "Local development debugging interface for Hono apps",
   "author": "Fiberplane<info@fiberplane.com>",
@@ -7,6 +7,7 @@
   "engines": {
     "node": ">=16.0.0"
   },
+  "types": "./src/lib/types.ts",
   "scripts": {
     "dev": "tsx watch src/index.node.ts",
     "db:generate": "drizzle-kit generate",
@@ -41,7 +42,6 @@
     "drizzle-orm": "^0.30.10",
     "drizzle-zod": "^0.5.1",
     "figlet": "^1.7.0",
-    "fpx-shared": "*",
     "hono": "^4.3.7",
     "minimatch": "^10.0.1",
     "octokit": "^4.0.2",
