+++ api/src/lib/types.ts
@@ -1,5 +1,6 @@
 // Only using for global types
 
+import { z } from "zod";
 import type { LibSQLDatabase } from "drizzle-orm/libsql";
 import type { WebSocket } from "ws";
 import type * as schema from "../db/schema.js";
@@ -14,3 +15,29 @@ export type Variables = {
   db: LibSQLDatabase<typeof schema>;
   wsConnections: Set<WebSocket>;
 };
+
+export type WsMessage = z.infer<typeof WsMessageSchema>;
+
+export const WsMessageSchema = z.discriminatedUnion("event", [
+  z.object({
+    event: z.literal("trace_created"),
+    // TODO: this should be an array of traces instead of the queryKeys to invalidate on the browser
+    payload: z.array(z.literal("mizuTraces")),
+  }),
+  z.object({
+    event: z.literal("connection_open"),
+    payload: z.object({
+      connectionId: z.string(),
+    }),
+  }),
+  z.object({
+    event: z.literal("request_incoming"),
+    payload: z.object({
+      headers: z.record(z.string()),
+      query: z.record(z.string()),
+      path: z.array(z.string()),
+      body: z.any(),
+      method: z.string(),
+    }),
+  }),
+]);
