--- shared/src/index.ts
@@ -1,1 +1,0 @@
-export { type WsMessage, WsMessageSchema } from "./ws.js";
